+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii 2 HTML to PDF conversion extension Change Log
 1.0.3 under development
 -----------------------
 
+- Enh #6: Options 'coverContent', 'headerHtmlContent' and 'footerHtmlContent' added tp `Wkhtmltopdf` (berosoboy, klimov-paul)
 - Enh #8: 'wkhtmltopdf' command composition improved ensuring options 'cover' and 'toc' do not utilize global ones (klimov-paul)